autoMinMaxSynth: add rangePad and t_forceUpdate triggered arg
rangePad allows you to specify a normalized padding to the displayed
range of min and max so that values at the min and max aren’t obscured
by the bottom of the view or the channel separator. Default 0.05.
t_forceUpdate allows you to force the min/max to update by triggering
the SendReply. This is used when autoMinMax_ is set so that the range
updates immediately. Previously a signal whose range was unchanging but
out of bounds wouldn’t trigger the min-max synth to send a reply to the
update the view.